Webpack confguration
This project is the final milestone of the the first part (Build Tools & Single Page Web Apps) of Udacity's Front End Web Development Advanced Nanodegree program. This project aims to build a web tool that allows users to run Natural Language Processing (NLP) on articles or blogs found on other websites. When a user submits a URL of an article, the web page then dispalys sentiment analysis returned from meaningcloud API, based on the contents of the article